--- a/frontend/android/build.gradle
+++ b/frontend/android/build.gradle
@@ -41,8 +41,8 @@ android {
     lintOptions {
         lintConfig file("${project.projectDir}/lint.xml")
 
-        xmlReport System.getenv("CI") == "true"
-        htmlReport System.getenv("CI") != "true"
+        xmlReport isCi
+        htmlReport !isCi
 
         xmlOutput file("lint-results.xml")
         htmlOutput file("lint-results.html")