--- a/frontendcomponent/androidtestcomponent/build.gradle
+++ b/frontendcomponent/androidtestcomponent/build.gradle
@@ -23,7 +23,7 @@ dependencies {
     }
 }
 repositories {
-    maven { url "https://maven-central-asia.storage-download.googleapis.com/repos/central/data/" }
+    if (!isCi) { maven { url "https://maven-central-asia.storage-download.googleapis.com/repos/central/data/" } }
     mavenCentral()
 }
 tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {