## Issue
- close #168

## Overview (Required)
- use Maven Central mirror following to https://storage-download.googleapis.com/maven-central/index.html

## Links
- https://storage-download.googleapis.com/maven-central/index.html

## Screenshot

There is no screenshot but build speed faster than before, I could not measure accurately though.

The following is synchronization time on my MacBook Air. I only measured once.

Before | After
:--: | :--:
synced successfully 10m 29s 199ms | synced successfully 8m 30s 38ms