@takahirom 2019-01-09
Since CI's server is overseas, how about judging whether it is CI server or not like this?

def isCi = System.getenv("CI") == "true"

```
        google()
        if (!isCi) { maven { url "https://maven-central-asia.storage-download.googleapis.com/repos/central/data/" } }
        jcenter()
```