@ma2gedev 2019-01-10
> Since CI's server is overseas, how about judging whether it is CI server or not like this?

Wow, I have not noticed about CI. That is a cool idea so I will check it later tonight in JST.