@ma2gedev 2019-01-10
I added `isCi` as extra property but please tell me if there is any other good way.